# next.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Job Board API",
    "version": "1.0.0",
    "description": "API for the AI Job Board application — browse job titles, search jobs, and submit applications."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session"
      }
    },
    "schemas": {
...
      "JobListing": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "external_id": { "type": "string" },
          "title": { "type": "string" },
          "company": { "type": "string" },
          "location": { "type": "string" },
          "job_type": { "type": "string" },
          "salary_range": { "type": "string" },
          "skills": { "type": "array", "items": { "type": "string" } },
          "description": { "type": "string" },
          "apply_url": { "type": "string" },
          "apply_email": { "type": "string" },
          "source": { "type": "string" },
          "expires_at": { "type": "string" },
          "created_at": { "type": "string" }
        }
      },
      "JobListingsPage": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/JobListing" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
      "JobListingCreateRequest": {
        "type": "object",
        "required": ["title", "company"],
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "company": { "type": "string", "minLength": 1 },
          "location": { "type": "string" },
          "job_type": { "type": "string" },
          "salary_range": { "type": "string" },
          "skills": { "type": "array", "items": { "type": "string" } },
          "description": { "type": "string" },
          "apply_url": { "type": "string" },
          "apply_email": { "type": "string" },
          "source": { "type": "string" },
          "expires_in_days": { "type": "integer", "minimum": 1, "maximum": 365 }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" }
        }
      }
    }
  },
  "paths": {
...
    "/apply": {
      "post": {
        "summary": "Submit a job application",
        "description": "Submits an application with resume upload. Accepts multipart/form-data.",
        "tags": ["Public"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "position", "type", "salary", "resume"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "description": "Applicant full name" },
                  "email": { "type": "string", "format": "email", "description": "Applicant email" },
                  "position": { "type": "string", "minLength": 1, "description": "Job position applied for" },
                  "type": { "type": "string", "minLength": 1, "description": "Job type (e.g. Full-time, Remote)" },
                  "salary": { "type": "string", "minLength": 1, "description": "Expected salary" },
                  "resume": { "type": "string", "format": "binary", "description": "Resume PDF file" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Application submitted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "429": { "description": "Too many requests" }
        }
      }
    },
...
    "/job-listings": {
      "get": {
        "summary": "Browse job listings with pagination",
        "tags": ["Public"],
        "parameters": [
          { "name": "search", "in": "query", "required": false, "description": "Search by title or company", "schema": { "type": "string" } },
          { "name": "location", "in": "query", "required": false, "description": "Filter by location", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "required": false, "description": "Filter by job type", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "required": false, "description": "Page number (default 1)", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "required": false, "description": "Results per page (default 10, max 50)", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated job listings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobListingsPage" }
              }
            }
          }
        }
      }
    },
...
    "/admin/job-listings": {
      "get": {
        "summary": "List all job listings (admin)",
        "tags": ["Admin - Job Listings"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Array of job listings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/JobListing" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Too many requests" }
        }
      },
      "post": {
        "summary": "Create a new job listing",
        "tags": ["Admin - Job Listings"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/JobListingCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job listing created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobListing" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Too many requests" }
        }
      },
      "delete": {
        "summary": "Delete a job listing",
        "tags": ["Admin - Job Listings"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "id", "in": "query", "required": true, "description": "Job listing ID", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Job listing deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": { "description": "Missing id parameter" },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Too many requests" }
        }
      }
    }
  },
...
}